Move more logic to audio utils
