+++ src/lib/audio/audio-util.js
@@ -1,4 +1,5 @@
-const computeRMS = function (samples) {
+const computeRMS = function (samples, scaling = 0.55) {
+    if (samples.length === 0) return 0;
     // Calculate RMS, adapted from https://github.com/Tonejs/Tone.js/blob/master/Tone/component/Meter.js#L88
     let sum = 0;
     for (let i = 0; i < samples.length; i++) {
@@ -6,13 +7,21 @@ const computeRMS = function (samples) {
         sum += Math.pow(sample, 2);
     }
     const rms = Math.sqrt(sum / samples.length);
-    // Scale it
-    const unity = 0.55;
-    const val = rms / unity;
-    // Scale the output curve
+    const val = rms / scaling;
     return Math.sqrt(val);
 };
 
+const computeChunkedRMS = function (samples, chunkSize = 1024) {
+    const sampleCount = samples.length;
+    const chunkLevels = [];
+    for (let i = 0; i < sampleCount; i += chunkSize) {
+        const maxIndex = Math.min(sampleCount - 1, i + chunkSize);
+        chunkLevels.push(computeRMS(samples.slice(i, maxIndex), true));
+    }
+    return chunkLevels;
+};
+
 export {
-    computeRMS
+    computeRMS,
+    computeChunkedRMS
 };
